migration scripts generate and repair (#132)
* fixd old migration file

* fixd old migration file

* add custom logic to alembic env

* remove exclude path

* generate new migration file for address_index and opensea tables

* fixed reorg column server default value

* Standardize column data type & Part of tables align with the online version

* add migration scripts for uniswap and merchant-moe daily table

* add migration scripts for merchant-moe table

* modify column name

---------

+++ indexer/modules/user_ops/models/user_operation_results.py
@@ -1,7 +1,7 @@
 from typing import Type
 
 from sqlalchemy import Column, Index
-from sqlalchemy.dialects.postgresql import BOOLEAN, BYTEA, INTEGER, NUMERIC, TEXT, TIMESTAMP, VARCHAR
+from sqlalchemy.dialects.postgresql import BIGINT, BOOLEAN, BYTEA, INTEGER, NUMERIC, TIMESTAMP, VARCHAR
 
 from common.models import HemeraModel, general_converter
 
@@ -29,7 +29,7 @@ class UserOperationResult(HemeraModel):
 
     transactions_hash = Column(BYTEA)
     transactions_index = Column(INTEGER)
-    block_number = Column(INTEGER)
+    block_number = Column(BIGINT)
     block_timestamp = Column(TIMESTAMP)
     bundler = Column(VARCHAR(42))
     start_log_index = Column(INTEGER)
